requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
st-aggrid==0.3.4.post3
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4 